requirements.txt
pandas
pdfplumber
gspread
google-auth
google-api-python-client
sentence-transformers
scikit-learn
numpy
python-dotenv
flask
flask-cors